Remove `sign` subroutine
As it's noticed, the mechanism of using `sign` to register elements can
be troublesome when multiple elements exist. The developer may forget to
register an element after creating it and setting up its connection. And
this mistake generally raises neither error nor warning.

After removing `sign`, the elements are added to a queue when created,
and generated into Verilog code when `write_top` is called.