Remove bit width check in `connect`
Bit width check can no longer be performed if we're introducing
parameterized design into the project. Parameters, especially those
passed in, cannot be calculated into an exact value during generation
time, making it impossible to compare two ports' widths.

Luckily, modern Verilog compilers have mostly equipped with the ability
to perform width check on compiling time. As long as the code is to be
compiled before a commit, mismatches hide nowhere to be a potential
problem.